Extract Codelab methods to their own file (data.dart)
