+++ lib/src/model/data.dart
@@ -1,150 +1,91 @@
-import 'dart:math';
+import 'package:cloud_firestore/cloud_firestore.dart';
 
-final List<String> locations = <String>[
-  'Albuquerque',
-  'Arlington',
-  'Atlanta',
-  'Austin',
-  'Baltimore',
-  'Boston',
-  'Charlotte',
-  'Chicago',
-  'Cleveland',
-  'Colorado Springs',
-  'Columbus',
-  'Dallas',
-  'Denver',
-  'Detroit',
-  'El Paso',
-  'Fort Worth',
-  'Fresno',
-  'Houston',
-  'Indianapolis',
-  'Jacksonville',
-  'Kansas City',
-  'Las Vegas',
-  'Long Island',
-  'Los Angeles',
-  'Louisville',
-  'Memphis',
-  'Mesa',
-  'Miami',
-  'Milwaukee',
-  'Nashville',
-  'New York',
-  'Oakland',
-  'Oklahoma',
-  'Omaha',
-  'Philadelphia',
-  'Phoenix',
-  'Portland',
-  'Raleigh',
-  'Sacramento',
-  'San Antonio',
-  'San Diego',
-  'San Francisco',
-  'San Jose',
-  'Seattle',
-  'Tucson',
-  'Tulsa',
-  'Virginia Beach',
-  'Washington',
-];
+import './filter.dart';
+import './restaurant.dart';
+import './review.dart';
 
-final List<String> cuisines = <String>[
-  'Brunch',
-  'Burgers',
-  'Coffee',
-  'Deli',
-  'Dim Sum',
-  'Indian',
-  'Italian',
-  'Mediterranean',
-  'Mexican',
-  'Pizza',
-  'Ramen',
-  'Sushi',
-];
+// This is the file that Codelab users will primarily work on.
 
-final List<String> _mockWords = <String>[
-  'Bar',
-  'Deli',
-  'Diner',
-  'Fire',
-  'Grill',
-  'Drive Thru',
-  'Place',
-  'Best',
-  'Spot',
-  'Trattoria',
-  'Steakhouse',
-  'Churrasco',
-  'Tavern',
-  'Cafe',
-  'Pop-up',
-  'Yummy',
-  'Belly',
-  'Snack',
-  'Fast',
-  'Turbo',
-  'Hyper',
-  'Prime',
-  'Eatin\'',
-];
-
-final Map<int, List<String>> _mockReviewsData = <int, List<String>>{
-  1: [
-    'Would never eat here again!',
-    'Such an awful place!',
-    'Not sure if they had a bad day off, but the food was very bad.'
-  ],
-  2: [
-    'Not my cup of tea.',
-    'Unlikely that we will ever come again.',
-    'Quite bad, but I\'ve had worse!'
-  ],
-  3: [
-    'Exactly okay :/',
-    'Unimpressed, but not disappointed!',
-    '3 estrellas y van que arden.'
-  ],
-  4: [
-    'Actually pretty good, would recommend!',
-    'I really like this place, I come quite often!',
-    'A great experience, as usual!'
-  ],
-  5: [
-    'This is my favorite place. Literally',
-    'This place is ALWAYS great!',
-    'I recommend this to all my friends and family!'
-  ],
-};
-
-final Random random = Random();
-
-String getMockReview(int rating) {
-  List reviews = _mockReviewsData[rating];
-  return reviews[random.nextInt(reviews.length)];
+Future<void> addRestaurant(Restaurant restaurant) async {
+  CollectionReference restaurants =
+      Firestore.instance.collection('restaurants');
+  return restaurants.add({
+    'avgRating': restaurant.rating,
+    'category': restaurant.cuisine,
+    'city': restaurant.location,
+    'name': restaurant.name,
+    'numRatings': restaurant.numRatings,
+    'photo': restaurant.imageUrl,
+    'price': restaurant.price,
+  });
 }
 
-String getMockName() {
-  int firstWord = random.nextInt(_mockWords.length);
-  int nextWord;
-  do {
-    nextWord = random.nextInt(_mockWords.length);
-  } while (firstWord == nextWord);
-  return '${_mockWords[firstWord]} ${_mockWords[nextWord]}';
+Stream<QuerySnapshot> loadAllRestaurants() {
+  return Firestore.instance
+      .collection('restaurants')
+      .orderBy('avgRating', descending: true)
+      .limit(50)
+      .snapshots();
 }
 
-String getMockCuisine() {
-  return cuisines[random.nextInt(cuisines.length)];
+Stream<QuerySnapshot> loadFilteredRestaurants(Filter filter) {
+  Query collection = Firestore.instance.collection('restaurants');
+  if (filter.cuisine != null) {
+    collection = collection.where('category', isEqualTo: filter.cuisine);
+  }
+  if (filter.location != null) {
+    collection = collection.where('city', isEqualTo: filter.location);
+  }
+  if (filter.price != null) {
+    collection = collection.where('price', isEqualTo: filter.price);
+  }
+  return collection
+      .orderBy(filter.sort ?? 'avgRating', descending: true)
+      .limit(50)
+      .snapshots();
 }
 
-String getMockCity() {
-  return locations[random.nextInt(locations.length)];
+List<Restaurant> getRestaurantsFromQuery(QuerySnapshot snapshot) {
+  return snapshot.documents.map((DocumentSnapshot doc) {
+    return Restaurant.fromSnapshot(doc);
+  }).toList();
 }
 
-String getMockImageUrl() {
-  int photoId = random.nextInt(21) + 1;
-  return 'https://storage.googleapis.com/firestorequickstarts.appspot.com/food_$photoId.png';
+Stream<DocumentSnapshot> getRestaurant(String restaurantId) {
+  return Firestore.instance
+      .collection('restaurants')
+      .document(restaurantId)
+      .snapshots();
+}
+
+Future<void> addReview(
+    {String restaurantId, String userId, String userName, Review review}) {
+  CollectionReference collection = Firestore.instance.collection('restaurants');
+  DocumentReference restaurant = collection.document(restaurantId);
+  DocumentReference newReview = restaurant.collection('ratings').document();
+
+  return Firestore.instance.runTransaction((Transaction transaction) {
+    return transaction
+        .get(restaurant)
+        .then((DocumentSnapshot freshRestaurantSnapshot) {
+      Restaurant freshRestaurant =
+          Restaurant.fromSnapshot(freshRestaurantSnapshot);
+      double newAverage =
+          ((freshRestaurant.numRatings * freshRestaurant.rating) +
+                  review.rating) /
+              (freshRestaurant.numRatings + 1);
+      transaction.update(restaurant, {
+        'numRatings': freshRestaurant.numRatings + 1,
+        'avgRating': newAverage,
+      });
+
+      return transaction.set(newReview, {
+        'rating': review.rating,
+        'text': review.text,
+        'userName': userName,
+        'timestamp': FieldValue.serverTimestamp(),
+        'userId': userId,
+      });
+    });
+  });
 }
