+++ lib/src/restaurant_page.dart
@@ -8,6 +8,7 @@ import 'package:flutter/material.dart';
 import 'package:sliver_fab/sliver_fab.dart';
 
 import './empty_list.dart';
+import './model/data.dart' as data;
 import './model/restaurant.dart';
 import './model/review.dart';
 import './restaurant_app_bar.dart';
@@ -32,11 +33,8 @@ class _FriendlyEatsRestaurantPageState
     extends State<FriendlyEatsRestaurantPage> {
   _FriendlyEatsRestaurantPageState({@required String restaurantId}) {
     FirebaseAuth.instance.signInAnonymously().then((AuthResult auth) {
-      _currentRestaurantSubscription = Firestore.instance
-          .collection('restaurants')
-          .document(restaurantId)
-          .snapshots()
-          .listen((DocumentSnapshot snap) {
+      _currentRestaurantSubscription =
+          data.getRestaurant(restaurantId).listen((DocumentSnapshot snap) {
         _currentReviewSubscription?.cancel();
         setState(() {
           _restaurant = Restaurant.fromSnapshot(snap);
@@ -73,38 +71,16 @@ class _FriendlyEatsRestaurantPageState
   List<Review> _reviews = <Review>[];
 
   Future<void> _addReview(String restaurantId, Review newReview) async {
-    CollectionReference collection =
-        Firestore.instance.collection('restaurants');
-    DocumentReference restaurant = collection.document(restaurantId);
-    DocumentReference review = restaurant.collection('ratings').document();
     String userId = await FirebaseAuth.instance
         .currentUser()
         .then((FirebaseUser user) => user.uid);
+    String userName = 'Anonymous (${kIsWeb ? "Web" : "Mobile"})';
 
-    return Firestore.instance.runTransaction((Transaction transaction) {
-      return transaction
-          .get(restaurant)
-          .then((DocumentSnapshot freshRestaurantSnapshot) {
-        Restaurant freshRestaurant =
-            Restaurant.fromSnapshot(freshRestaurantSnapshot);
-        double newAverage =
-            ((freshRestaurant.numRatings * freshRestaurant.rating) +
-                    newReview.rating) /
-                (freshRestaurant.numRatings + 1);
-        transaction.update(restaurant, {
-          'numRatings': freshRestaurant.numRatings + 1,
-          'avgRating': newAverage,
-        });
-
-        return transaction.set(review, {
-          'rating': newReview.rating,
-          'text': newReview.text,
-          'userName': 'Anonymous (${kIsWeb ? "Web" : "Mobile"})',
-          'timestamp': FieldValue.serverTimestamp(),
-          'userId': userId,
-        });
-      });
-    });
+    return data.addReview(
+        restaurantId: restaurantId,
+        userId: userId,
+        userName: userName,
+        review: newReview);
   }
 
   void _onCreateReviewPressed(BuildContext context) async {
